+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## HEAD
+
+* `presign_endpoint` – Add `Shrine.presign_response` for handling presigns inside a custom controller (@janko)
+
+* `upload_endpoint` – Add `Shrine.upload_response` for handling uploads inside a custom controller (@janko)
+
 ## 2.17.1 (2019-05-15)
 
 * `rack_file` – Fix overriden `Attacher#assign` not accepting second argument (@janko)